Fix compile issue with theme colour asserts
